requirements.txt
fastapi
python-jobspy
pinecone
sentence-transformers
sqlalchemy
sqlalchemy-libsql
python-dotenv
numpy
requests
beautifulsoup4
pydantic
uvicorn